qualified expression with cast

+++ xd/xdpgsql/database.cpp
@@ -683,7 +683,7 @@ std::wstring makeStreamReference(xd::IStream* stream)
 {
     wchar_t buf[80];
     if (sizeof(void*) == 8)
-        swprintf(buf, 80, L"streamptr://%llX", (void*)stream);
+        swprintf(buf, 80, L"streamptr://%llX", (long long)(void*)stream);
          else
         swprintf(buf, 80, L"streamptr://%X", (void*)stream);
     return buf;
